src/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==23.1.0
audioread==3.0.0
billiard==3.6.4.0
blinker==1.6.2
brotlipy==0.7.0
cachetools==5.3.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
colorama==0.4.6
configobj==5.0.8
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
databricks-cli==0.17.7
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.1
distro==1.8.0
docker==6.1.2
dpath==2.1.5
dulwich==0.21.5
dvc==2.57.2
dvc-data==0.51.0
dvc-gdrive==2.19.2
dvc-http==2.30.2
dvc-objects==0.22.0
dvc-render==0.5.2
dvc-studio-client==0.9.2
dvc-task==0.2.1
entrypoints==0.4
filelock==3.9.0
Flask==2.3.2
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
funcy==2.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
gmpy2==2.1.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
grandalf==0.8
greenlet==2.0.2
grpcio==1.54.2
gunicorn==20.1.0
httplib2==0.22.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
iterative-telemetry==0.0.8
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
lazy_loader==0.2
librosa==0.10.0.post2
lightning-utilities==0.8.0
llvmlite==0.40.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.1
mdurl==0.1.2
mlflow==2.3.2
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
nanotime==0.5.2
networkx==2.8.4
numba==0.57.0
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
orjson==3.8.12
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.4.0
pip==23.0.1
platformdirs==3.5.1
pooch==1.6.0
prompt-toolkit==3.0.38
protobuf==4.23.1
psutil==5.9.5
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyDeprecate==0.3.1
pydot==1.4.2
PyDrive2==1.15.3
pygit2==1.12.1
Pygments==2.15.1
pygtrie==2.5.0
PyJWT==2.7.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.5.6
pytz==2023.3
PyYAML==6.0
querystring-parser==1.2.4
requests==2.29.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
scmrepo==1.0.3
setuptools==66.0.0
shortuuid==1.0.11
shtab==1.6.1
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.5
SQLAlchemy==2.0.14
sqlparse==0.4.4
sqltrie==0.3.1
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
tomlkit==0.11.8
torch==1.13.1
torchvision==0.14.1
torchaudio
torchdata
torchgeometry
torchmetrics==0.11.4
torchsummary==1.5.1
torchtext==0.6.0
tqdm==4.65.0
scikit-image
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.3.4
wheel==0.38.4
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.15.0
